nexus-ai-addon/requirements.txt
# Web framework
fastapi==0.105.0
uvicorn==0.23.2
pydantic==2.5.1
starlette==0.27.0
aiohttp==3.9.1
requests==2.31.0
jinja2==3.1.2
python-multipart==0.0.6
websockets==12.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# AI and Vector Storage
openai==1.3.7
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.0.350
numpy==1.26.2

# Calendar Support
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.109.0

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1
python-slugify==8.0.1
pytz==2023.3.post1
emoji==2.8.0
